[DC-68] feature: design interfaces for wizard page controllers (#12184)
* Changed the wizard related QList<QSslCertificate> to QSet, as this jives with the final type in the account
added a bare bones start of a controller for the first page of the wizard which will evaluate a url.

* added the new urlpagecontroller to the correct CMakeLists.txt

* added missing namespace
added more plans/notes on the UrlPageController and NewAccountWizardController
added new, mostly empty OAuthPageController

* added an abstract WizardPageValidator to allow exposing the controllers' validate methods to the wizard while keeping the other functionality in hidden/inaccessible.
added more notes for impl plans.
this is as far as I can take the design without starting an impl so I'm calling this done.