[DC-70] feature: new account wizard base implementation including first page (#12189)
* implemented basics for the main new account wizard gui - a QWizard subclass

* added some hacks to allow showing the new wizard.
Important notes: you can only show it using the "add account" button for now, and you can only cancel out of it.
#define USE_NEW_ACCOUNT_WIZARD appears in the SettingsDialog - this will eventually go away but you can comment it out to restore the normal add account behavior from the button.
why you would do that on this branch I don't know, but you could.

* added all the pages and controllers
implemented the basic button text and presence logic

* added basic page content for the first page
the impl is super messy but "works" as a first pass
the url checking is horrid - we really need a (very custom) url validator on this field asap.
it bahaves close to how it used to, so that's something.
the validation does complete with a valid server
I think this will be an excellent test for a stack of three dialogs, two shown via blocking, normal exec (the main window, the wizard, and the tls error dialog on top of all that) but have not tried it yet.

* fixed the wizard size and improved the look of the first wizard page + added the main app logo

* added new footer logo (only used in kiteworks branding, for now)

* taking out a trash assert that resulted in several hours of confusion.

* added a new theme function to get a placeholder url for the first page of the wizard. note this will not be visible until the kiteworks theme is updated (later)
also fiddled around with the layout of the first wizard page.
updated placeholder text on the "empty" pages of the wizard.

* added a new "dark" logo to support kiteworks when dark theme is in play. In general, we only try to use this alternate logo when no wizard background color is set, and the theme provides the logo. otherwise just use the regular logo.
also successfully implemented the bg/text override colors in the wizard using QPalette instead of a style sheet
this update allowed me to successfully set the error text color to red. Tested this red text with kiteworks (light and dark mode) + "unbranded" oc build (which uses dark blue background for the wizard) = all good so far.

* added handling for theme provided server url
also declared injected deps as qpointers - this is not really necessary since all components have shared lifetime but doesn't hurt and is best practice.

* added the AccessManager::reset functionality
removed/didn't implement any support for http scheme
added really key todo's relative to fallback to oc10 style auth in oauth.cpp
removed pointless includes in cookiejar.cpp and added a todo to review if this class is useful at all.

* This completes the first pass for the url page in the wizard.
For now the results are returned by signal, and I think I prefer this as we should handle the data before the next page is actually displayed.
At the moment the values collected from the first page are simply inserted into the main wizard controller's model. This still needs more work but the main concept is there now.

* changed the ResolveUrlAdapter to take a QWidget argument to parent the tls error dialog. Without this change, the dialog is parented to the main window, which doesn't work 100% on linux.
also changed the wizard options so that the cancel button is no longer hidden on mac
got rid of the placeholder text for the error message

* hopefully this fixes the focus problem with the url field on mac

* forgot to remove the hide url field if the url is provided by the theme. this needs further ux review. for now I think showing but disabling the field is fine.

* removing the switch that enables the new account wizard via the add account button. For testing local builds with the new wizard, uncomment the #define USE_NEW_ACCOUNT_WIZARD

* final cleanup

* turn off the dialog again

* review feedback

* replacing accidentally removed setFocus for mac