[DC-78] page three of new account wizard (#12212)
* basic layout of the success/third page of the wizard
I have a kinda ok result on the look of the AdvancedSettings button. I set it to NoFocus as it doesn't show focus properly in the first place, it steals focus from the button bar, AND it's also supposed to be a "stealth" option so leaving it with a click only trigger. Need to talk to Erik about this in terms of accessibility. I don't like it but if there is no blocker with accessibility ok.

* implemented specific button layouts for each page
connected the "advanced settings" buttons to advance the page if either one is clicked (only one will be in the final impl, we need to test both especially for accessibility)

* fixed a visible typo in folderman
updated the docs on NewAccountWizardController

* fix squish