[DC-86] refactor credentials related impls (#12253)
* primary changes are to credentials manager. I wanted to tighten it up, moved the global functions into the class as members, got rid of public accessor for account and also removed the dep on CredentialManager from the CredentialJob. Also did some cleanup to clarify when a key is supposed to be scoped or not.
Finally, since it's perfectly legitimate that the credentials manger may return a nullptr for the get job, updated callers to deal with that possibility.
For future improvement we should refactor the remove and set functions to return nullptr instead of relying on the asserts currently there.

* fix non-windows builds

* wrong class last time

* this fixes problem when re-login is attempted, as the get job comes back null but needed more handling

* eliminated all but one friend relationship in the CredentialManager.

* new credentials class that handles oauth only. Naming it credentials as we have no others in play now nor in forseeable future.
so far the credentials are a copy of the HttpCredentials which I will refine as I go along.
At this point I have removed the auth type and checks that determine of oauth is active. It is always active now so the "else"s are gone and members refer to oauth appropriate values
also started adding notes for the future refinement

* small changes to general creds handling
big overhaul of new Credentials class
lots of notes about changes that should happen with DC-112

* hopefully fixing tests

* brainfart

* new classes to cover the "askFromUser" functionality in the creds.
the widget is pretty much done, the controller still to be implemented.

* added the buttons to the gui and updated text

* fleshed out the controller and made some corresponding changes to the widget

* the re-auth is now working!
and it looks pretty good too.

* update selector object of copy url to clipboard

* cleanup - also tried to remove includes to httpcredentials and httpcredentialsgui
there is one "used" httpcredentials case but this is highly questionable and needs a deeper look to determine if it needs to be updated or just removed. I will do that in DC-112

* updated accessibility on the new log in widget

---------

Co-authored-by: Prajwol Amatya <prajwolamatya11@gmail.com>