[DC-128] refactor concerning areas of credentials in relation to account (#12289)
* updated remaining todo's for dc-112 to new story id 128

* updated the abstract creds to take both an account and parent
deleted several outdated functions from the base and subs, most notably the creds->setAccount (it should only come from the ctr)
updated the tests to use the new scheme

* fixed the tests.

* get rid of the "save credentials" switches on account save routines. the creds persist to the keychain as needed so these args on account related save are just noise.

* mostly test refactoring to cover the new reqs for creating account before creds (as the creds take account in ctr)
I also found the source of almost endless asserts on the account->davUser which should not be empty. That was accessed in debug output BEFORE the value was set :/ davUser is moving to the acct ctr asap but that was such a sneaky failure.

* tests finally pass

* start dealing with the setDavUser

* the account url and davUser are now immutable after construction
also added a new todo for getting rid of more oc10 stuff

* Finally updated the Account::id to reflect what it actually is: the account group index which is used to write the settings
also removed the deprecation flag as we need this into the future.
also renamed the associated functions in the account manager to reflect the index based nature of the value

* final cleanup

* fixed failing testconnectionvalidator

* fixed major bug that the credentials were not persisted after creating a new account, which caused a re-auth request when restarting the app (only the first time after creating the account. after reauth everything is cool)
now it works correctly so the account is created, and properly reloaded with the existing creds on restart.